requirements.txt
python-telegram-bot==20.3
Flask==2.3.3
psycopg2-binary==2.9.9
python-dotenv